+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## [0.4.0]
+### Changed
+- Now looks for `:keenex` config block
+- Removed start_link functions
+
 ## [0.3.0]
 ### Changed
 - API now more closely aligned to existing keen.io libraries. The `Keenex` module now has all functions for interacting with keen.io's API. `Keenex.Events` and `Keenex.EventCollections` modules are removed